Import many passwords

+++ app/Http/Repositories/Password/PasswordRepository.php
@@ -18,6 +18,11 @@ class PasswordRepository {
         return $password;
     }
 
+    private function retrievePasswordFromName(string $name) : Password | null {
+        $password = Auth::user()->passwords()->where('name', $name)->first();
+        return $password;
+    }
+
     public function retrieveList() {
         return Auth::user()->passwords;
     }
@@ -43,17 +48,22 @@ class PasswordRepository {
     private function removeFolderIfNeeded(int $folderId) : void {
         $folder = Auth::user()->folders()
             ->where([ 'id' => $folderId, 'model' => 'passwords' ])->first();
-        if ($folder->passwords->count() === 0) {
+        if (!is_null($folder) && $folder->passwords->count() === 0) {
             $folder->delete();
         }
     }
 
-    public function delete(int $passwordId) : void {
+    public function delete(int $passwordId) : bool {
         $model = $this->retrievePasswordFromId($passwordId);
         if (!$model) throw new PasswordNotFoundException();
         $folder = $model->folder;
         $model->delete();
-        if (!is_null($folder)) $this->removeFolderIfNeeded($folder->id);
+        if (!is_null($folder)) {
+            $this->removeFolderIfNeeded($folder->id);
+            return true;
+        }
+
+        return false;
     }
 
     private function createFolder(string $name) : Folder {
@@ -66,6 +76,41 @@ class PasswordRepository {
         $folder->name = strtoupper($name);
         $folder->save();
         return $folder;
+    }
+
+    /*
+      $passwords = PasswordModel[]
+    */
+    public function createMany(array $passwords) : array {
+
+        $createdCount = 0;
+        $updatedCount = 0;
+        $createdModels = [];
+
+        $names = array_map(function ($password) {
+            return $password->name;
+        }, $passwords);
+
+        $passwordModals = Auth::user()->passwords()->whereIn('name', $names)->get();
+
+        foreach ($passwords as $password) {
+            $model = $passwordModals->where('name', $password->name)->first();
+            $createdModel = array_key_exists($password->name, $createdModels) ? $createdModels[$password->name] : null;
+
+            if (is_null($model) && is_null($createdModel)) {
+                $model = $this->create($password);
+                $createdModels[$password->name] = $model;
+                $createdCount ++;
+                continue;
+            }
+
+            $model = $model ?? $createdModel;
+            $password->setId($model->id);
+            $model = $this->update($password, $model);
+            $updatedCount ++;
+        }
+
+        return [ 'created' => $createdCount, 'updated' => $updatedCount ];
     }
 
     public function create(PasswordModel $password) : Password {
@@ -96,8 +141,8 @@ class PasswordRepository {
         return $model;
     }
 
-    public function update(PasswordModel $password) : Password {
-        $model = $this->retrievePasswordFromId($password->id);
+    public function update(PasswordModel $password, Password $model = null) : Password {
+        $model = is_null($model) ? $this->retrievePasswordFromId($password->id) : $model;
 
         if (is_null($password) || is_null($password->id)) throw new PasswordNotFoundException();
